requirements.txt
streamlit
python-dotenv
groq
httpx
PyPDF2
python-docx
requests 